To legacy opticalFlow tests
